+++ apps/forms/scaffolds/playground-canvas/playground.tsx
@@ -16,8 +16,6 @@ import {
   CanvasOverlay,
   standaloneDocumentReducer,
   initDocumentEditorState,
-  useEventTarget,
-  CursorMode,
 } from "@/builder";
 import {
   Select,
@@ -32,19 +30,7 @@ import { GridaLogo } from "@/components/grida-logo";
 import { DevtoolsPanel } from "@/builder/devtools";
 import { FontFamilyListProvider } from "@/scaffolds/sidecontrol/controls/font-family";
 import { useGoogleFontsList } from "@/builder/google.fonts";
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
-import {
-  BorderSolidIcon,
-  BoxIcon,
-  CircleIcon,
-  CursorArrowIcon,
-  DownloadIcon,
-  FigmaLogoIcon,
-  FileIcon,
-  FrameIcon,
-  ImageIcon,
-  TextIcon,
-} from "@radix-ui/react-icons";
+import { DownloadIcon, FigmaLogoIcon, FileIcon } from "@radix-ui/react-icons";
 import KeyboardInputOverlay from "@/builder/devtools/keyboard-input-overlay";
 import {
   DropdownMenu,
@@ -61,6 +47,7 @@ import { v4 } from "uuid";
 import { grida } from "@/grida";
 import { HelpFab } from "@/scaffolds/help/editor-help-fab";
 import { Badge } from "@/components/ui/badge";
+import { PlaygroundToolbar } from "./toolbar";
 
 export default function CanvasPlayground() {
   const [exampleid, setExampleId] = useState<string>("helloworld.grida");
@@ -210,7 +197,7 @@ export default function CanvasPlayground() {
                 </div>
               </div>
               <div className="absolute bottom-20 left-0 right-0 flex items-center justify-center z-50 pointer-events-none">
-                <Toolbar />
+                <PlaygroundToolbar />
               </div>
               <div className="fixed bottom-20 left-10 flex items-center justify-center z-50 pointer-events-none">
                 <KeyboardInputOverlay />
@@ -268,74 +255,4 @@ function ExampleSwitch({
       </SelectContent>
     </Select>
   );
-}
-
-function Toolbar() {
-  const { setCursorMode, cursor_mode } = useEventTarget();
-
-  return (
-    <div className="rounded-full flex gap-4 border bg-background shadow px-4 py-2 pointer-events-auto">
-      <ToggleGroup
-        onValueChange={(v) => {
-          setCursorMode(
-            v
-              ? toolbar_value_to_cursormode(v as ToolbarToolType)
-              : { type: "cursor" }
-          );
-        }}
-        value={cursormode_to_toolbar_value(cursor_mode)}
-        defaultValue="cursor"
-        type="single"
-      >
-        <ToggleGroupItem value={"cursor" satisfies ToolbarToolType}>
-          <CursorArrowIcon />
-        </ToggleGroupItem>
-        <ToggleGroupItem value={"container" satisfies ToolbarToolType}>
-          <FrameIcon />
-        </ToggleGroupItem>
-        <ToggleGroupItem value={"text" satisfies ToolbarToolType}>
-          <TextIcon />
-        </ToggleGroupItem>
-        <ToggleGroupItem value={"rectangle" satisfies ToolbarToolType}>
-          <BoxIcon />
-        </ToggleGroupItem>
-        <ToggleGroupItem value={"ellipse" satisfies ToolbarToolType}>
-          <CircleIcon />
-        </ToggleGroupItem>
-        <ToggleGroupItem value={"line" satisfies ToolbarToolType}>
-          <BorderSolidIcon />
-        </ToggleGroupItem>
-        <ToggleGroupItem value={"image" satisfies ToolbarToolType}>
-          <ImageIcon />
-        </ToggleGroupItem>
-      </ToggleGroup>
-    </div>
-  );
-}
-
-type ToolbarToolType =
-  | "cursor"
-  | "rectangle"
-  | "ellipse"
-  | "text"
-  | "container"
-  | "image"
-  | "line";
-
-function cursormode_to_toolbar_value(cm: CursorMode): ToolbarToolType {
-  switch (cm.type) {
-    case "cursor":
-      return "cursor";
-    case "insert":
-      return cm.node;
-  }
-}
-
-function toolbar_value_to_cursormode(tt: ToolbarToolType): CursorMode {
-  switch (tt) {
-    case "cursor":
-      return { type: "cursor" };
-    default:
-      return { type: "insert", node: tt };
-  }
 }
